+++ aeneas/cwave/cwave_setup.py
@@ -50,7 +50,7 @@ CMODULE = Extension(
 
 setup(
     name="cwave",
-    version="1.5.2",
+    version="1.6.0",
     description="""
     Python C Extension for for reading WAVE files.
     """,